refactor(tests): repository-only 契約テストを tests/repo/ へ移す (#3043)
`youtube_automation` を import しない静的契約テスト 55 件を `tests/repo/` へ移す。
src レイヤに対応先が存在しない群で、`docs/architecture/tests-layout.md` の鏡像規則の
適用対象外となるため先に切り離す。

対象は `grep -L "youtube_automation" tests/test_*.py` の出力と完全に一致させ、
実装者の裁量を挟まない（#3084 が追加した `test_takt_workflow_contract.py` を含む）。
移動に伴う参照追従は CI の pytest target、docs、テスト間 import に限定した。

`tests/conftest.py` の lane 登録は basename ベースのため追従不要（#3042）。
`.takt/facets/**` は #3084 の 0.55.1 向け再設計で当該記述ごと消えたため対象外。
`docs/audits/2026-05-skill-md-audit.md` の旧パス参照は当時の作業を記録した歴史文書
なので更新しない。
