+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `refactor(takt)`: takt 0.55.1 の builtin workflow / step fragment を基準に `.takt/` の feature・fix・docs・maintenance・audit 系 workflow を再設計した。公開 lane の目的を維持しつつ、structured fail-closed gate、有限 loop monitor、CI 同等ゲート、外部副作用を起こさない spillover へ統一し、旧 callable・facet・schema と内部構造固定テストを利用者向け契約テストへ置換した。
+- `refactor(tests)`: repository-only contract testsを `tests/repo/` へ移し、CI・active documentation・workflow facet の参照を canonical path に追従させた（#3043）。
 - `chore(takt)`: `.takt/config.yaml` に `auto_requeue_max_attempts: 0` を追加し、失敗タスクの自動再実行を無効化した。yt-auto-* の workflow は intake / plan / diagnose で「着手すべきでない」と判断したとき設計どおり ABORT するが、auto_requeue は同じ order.md のまま回し直すため判断が覆らず試行回数とトークンだけを消費していた。あわせて `runAllTasks` 起動時の `requeueExistingFailedTasks` も止まり、過去の failed タスクが `takt run` のたびに掘り起こされることがなくなる。再実行は原因を解消してから明示的に積み直す（#3080）。
 - `refactor(tests)`: pytest lane registry の slow node 登録を module basename と test 識別子で表現し、`tests/` 以下の module 配置に依存しない lane 判定と再帰的な module 検索を可能にした（#3042）。
 - `refactor(tests)`: `tests.helpers.paths` に repository root、`tests/`、fixtures の共通 path helper を追加し、テストからの `Path(__file__)` 直接遡上を禁止する契約テストと tests 配置規約を導入した（#3041）。
