+++ tests/repo/test_analytics_analyze_skill_contract.py
