+++ tests/repo/test_analytics_collect_depth_validator.py
