+++ tests/repo/test_analytics_revenue_skill_contract.py
