+++ tests/repo/test_channel_new_ttp_duration.py
