+++ tests/repo/test_extension_oxlint_contract.py
