+++ tests/repo/test_extension_package_manager_contract.py
