+++ tests/repo/test_extensions_shared_ui_contract.py
