+++ tests/repo/test_market_research_skill_contract.py
