+++ tests/repo/test_shadcn_skill_contract.py
