+++ tests/repo/test_skill_cost_documentation.py
