+++ tests/repo/test_suno_skill_doc.py
