+++ tests/repo/test_thumbnail_codex_image_skill.py
