+++ tests/repo/test_thumbnail_iterate.py
