+++ tests/repo/test_upgrade_guide_command_guard.py
