+++ tests/repo/test_verify_extensions_script.py
